refactor(#742): relocate identifier-callee dispatch to call-identifier.ts
Wave B chunk 1, step 2. Move compileIdentifierCall (the identifier-callee
dispatch family — node:fs globals, inline global builtins, direct
named-function calls) out of calls.ts into a new sibling module
src/codegen/expressions/call-identifier.ts. The 14 calls.ts internals it needs
are exported from calls.ts; calls.ts imports compileIdentifierCall back (a lazy
cycle matching the existing calls.ts <-> calls-closures/new-super cycles).

calls.ts: 19,435 -> 17,441 LOC. compileCallExpression: ~13,371 -> ~11,388 LOC.
New call-identifier.ts: 2,105 LOC.

Pure relocation — emitted Wasm byte-identical (prove-emit-identity IDENTICAL,
39/39 gc/standalone/wasi). tsc 0 errors. tests/issue-742.test.ts extended with
wasm=JS cases for the moved paths (parseInt family, Array(...), direct/recursive
named calls). The two *-allow keys in the issue frontmatter sanction the new
file the change-scoped loc-budget/coercion-sites gates flag (both net-zero
across the tree — a relocation, not new code).

✓
